+++ web-wallet/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add memo field to send flow [#3112]
+- Add "VITE_SYNC_INTERVAL" ENV variable [#3403]
 
 ### Changed
 
@@ -565,6 +566,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#3362]: https://github.com/dusk-network/rusk/issues/3362
 [#3381]: https://github.com/dusk-network/rusk/issues/3381
 [#3387]: https://github.com/dusk-network/rusk/issues/3387
+[#3403]: https://github.com/dusk-network/rusk/issues/3403
 [#3420]: https://github.com/dusk-network/rusk/issues/3420
 
 <!-- VERSIONS -->